pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelselector"
version = "0.1.0"
description = "Tkinter GUI for manually selecting pixel coordinates from an image"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = ["Pillow"]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "bandit",
    "vulture",
    "radon",
    "pytest",
    "pytest-cov",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=pixelselector --cov-report=term-missing --cov-report=xml --cov-fail-under=80"

[tool.coverage.run]
source_pkgs = ["pixelselector"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80

[tool.bandit]
targets = ["pixelselector.py"]
skips = []

[tool.setuptools]
py-modules = ["pixelselector"]
